requirements.txt
Flask~=3.1.2
httpx~=0.28.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.2.1
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
flask[async]
toml
pygments
pillow
mdurl
markdown-it-py
rich
./sb2gs